requirements.txt
absl-py==1.4.0
adabelief-tf==0.2.1
efficientnet==1.1.1
gpu-control==1.0.0
hyperas==0.4.1
imbalanced-learn==0.11.0
keras==2.10.0
keras-adabound==0.6.0
keras-efficientnet-v2==1.2.2
keras-gradient-noise==0.11
loguru==0.5.3
matplotlib==3.7.2
model-profiler==1.1.8
numpy==1.25.1
opencv-python==4.8.0.74
pandas==2.0.3
Pillow==10.1.0
py-cpuinfo==9.0.0
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
tensorflow==2.10.1
tensorflow-addons==0.22.0
tqdm==4.66.1
